Update utilsFS.cpp
i tried to put the option for loading LittleFS or other file systems in it for loading LittleFS: MJPEG2SD.htm common.js Auxil.htm from them. but i failed! it is going to be faster if we load pages from LittleFS and save files in SD card. we can even copy them into LittleFS if they dont exist.